added explicit cast, required by some compiler versions, namely on s390 arch
git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk/qpid@1213485 13f79535-47bb-0310-9956-ffa450edef68

+++ cpp/src/qpid/broker/MessageGroupManager.cpp
@@ -280,7 +280,7 @@ void MessageGroupManager::query(qpid::types::Variant::Map& status) const
          g != messageGroups.end(); ++g) {
         qpid::types::Variant::Map info;
         info[GROUP_ID_KEY] = g->first;
-        info[GROUP_MSG_COUNT] = g->second.members.size();
+        info[GROUP_MSG_COUNT] = (uint64_t) g->second.members.size();
         // set the timestamp to the arrival timestamp of the oldest (HEAD) message, if present
         info[GROUP_TIMESTAMP] = 0;
         if (g->second.members.size() != 0) {
